CORE-11534: Rename package
