+++ components/virtual-node/virtual-node-management-sender/src/main/java/net/corda/virtualnode/rest/common/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.virtualnode.rest.common;
+
+import org.osgi.annotation.bundle.Export;
