--- components/virtual-node/virtual-node-management-sender/src/main/java/net/corda/virtualnode/rpcops/common/package-info.java
@@ -1,4 +1,0 @@
-@Export
-package net.corda.virtualnode.rpcops.common;
-
-import org.osgi.annotation.bundle.Export;
