--- components/virtual-node/virtual-node-management-sender/src/main/kotlin/net/corda/virtualnode/rpcops/common/VirtualNodeSenderFactory.kt
@@ -1,11 +1,0 @@
-package net.corda.virtualnode.rpcops.common
-
-import net.corda.libs.configuration.SmartConfig
-import net.corda.messaging.api.publisher.config.PublisherConfig
-import java.time.Duration
-
-interface VirtualNodeSenderFactory {
-    fun createSender(
-        timeout: Duration, messagingConfig: SmartConfig, asyncPublisherConfig: PublisherConfig
-    ): VirtualNodeSender
-}
