+++ components/virtual-node/virtual-node-rest-service-impl/src/main/kotlin/net/corda/virtualnode/rest/impl/Constants.kt
@@ -1,0 +1,4 @@
+package net.corda.virtualnode.rest.impl
+
+internal const val GROUP_NAME = "virtual.node.management"
+internal const val CLIENT_NAME_HTTP = "virtual.node.manager.http"
