+++ components/virtual-node/virtual-node-rest-service-impl/src/main/kotlin/net/corda/virtualnode/rest/impl/v1/UnknownResponseTypeException.kt
@@ -1,0 +1,3 @@
+package net.corda.virtualnode.rest.impl.v1
+
+class UnknownResponseTypeException(type: String) : Exception("Encountered a response of unknown type: $type")
