--- components/virtual-node/virtual-node-rest-service-impl/src/main/kotlin/net/corda/virtualnode/rpcops/impl/Constants.kt
@@ -1,4 +1,0 @@
-package net.corda.virtualnode.rpcops.impl
-
-internal const val GROUP_NAME = "virtual.node.management"
-internal const val CLIENT_NAME_HTTP = "virtual.node.manager.http"
