--- components/virtual-node/virtual-node-rest-service-impl/src/main/kotlin/net/corda/virtualnode/rpcops/impl/v1/UnknownResponseTypeException.kt
@@ -1,3 +1,0 @@
-package net.corda.virtualnode.rpcops.impl.v1
-
-class UnknownResponseTypeException(type: String) : Exception("Encountered a response of unknown type: $type")
