+++ components/virtual-node/virtual-node-rpcops-maintenance-impl/src/main/kotlin/net/corda/libs/virtualnode/maintenance/rpcops/impl/v1/VirtualNodeMaintenanceRestResourceImpl.kt
@@ -33,8 +33,8 @@ import net.corda.schema.configuration.ConfigKeys
 import net.corda.utilities.debug
 import net.corda.utilities.time.UTCClock
 import net.corda.v5.base.exceptions.CordaRuntimeException
-import net.corda.virtualnode.rpcops.common.VirtualNodeSender
-import net.corda.virtualnode.rpcops.common.VirtualNodeSenderFactory
+import net.corda.virtualnode.rest.common.VirtualNodeSender
+import net.corda.virtualnode.rest.common.VirtualNodeSenderFactory
 import org.osgi.service.component.annotations.Activate
 import org.osgi.service.component.annotations.Component
 import org.osgi.service.component.annotations.Reference
